backend/requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
mysql-connector-python==8.3.0

# CORS
Flask-CORS==4.0.0

# JWT Authentication
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0

# Password Hashing
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Rate Limiting
Flask-Limiter==3.5.1

# Input Validation
marshmallow==3.20.1

# Email
Flask-Mail==0.9.1

# WSGI Server
gunicorn==21.2.0

# File Handling
python-magic==0.4.27

# Logging
colorlog==6.8.0

# Utilities
python-dateutil==2.8.2
